fix(datalayer): 修复 MigrationEngine 类型不匹配编译错误
schema.version 在 SQLDelight 2.x 为 Long，原 when 分支用 </> 比较（有运算符重载）
掩盖了 Int/Long 混用；新增的 == 分支无重载导致 compileReleaseKotlin 失败。
统一为 current.toLong() == target。

+++ app/src/main/java/com/R/codecore/datalayer/migration/MigrationEngine.kt
@@ -41,7 +41,7 @@ class MigrationEngine(private val pathProvider: DatabasePathProvider) {
                     .sortedBy { it.from }
                     .forEach { it.block(driver) }
             }
-            current == target -> {
+            current.toLong() == target -> {
                 // 显式 no-op。⚠️ 版本相等 ≠ 表结构一致：v0.5.0-rc1 事故中，表结构演进（+14 张表）
                 // 未伴随版本递增（彼时 0 个 .sqm，version 恒为 1），本分支静默空转导致
                 // 从 v0.4.0 升级的设备永远缺表、首查即崩（no such table: remote_connections）。
